Move fixtures matrix and display in specific modules.
